photowakeup/requirements.txt
absl-py==0.10.0
astunparse==1.6.3
automium==0.2.6
automium-web==0.1.1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
chumpy==0.70
click==7.1.2
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
Flask==1.1.2
future==0.18.2
gast==0.3.3
google-auth==1.21.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
legacy==0.1.6
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.0.3
networkx==2.4
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.4.0.42
opt-einsum==3.3.0
Pillow==7.2.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.15.0
scipy==1.4.1
six==1.15.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
tf-slim==1.1.0
torch==1.4.0
torchvision==0.5.0
tqdm==4.50.2
trimesh==3.8.11
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
zipp==1.2.0
